[emacs] unset selected window on window switch
If you update modeline on exiting window it is considered active by
modeline. This doesn't allow to inactivate segments that due to
optimisation are not updated eagerly.

This particular commit fixes the problem with vcs modeline segment being
active in inactive window. Others will be also fixed by updating them on
nucleus-exit-window-hook.