[notify] partial support for macOS
Still, enabling/disabling doesn't work as I could not find any good and reliable
way to temporarily disable notifications. It's possible to do some 'default
writes', but NotificationCenter doesn't catch the changes and on reload it just
resets everything. You can completely kill the center, but it's also not that
good as for me.