[emacs] navigation and completion in proto files
Basically, thanks to `counsel-etags`, `company-ctags` and some glue code, I have
navigation and completion in protobuf files. It works not only in the scope of
the current project I work right now, but also support navigation and completion
of entities defined in external libraries.

Completion is based on the ctags and ddabrev, so it's not context aware.

Right now management of external libraries is hard-coded (see
`+proto|install-dependenices`), but it's possible to abstract over this idea and
integrate with package managers to automatically support dependencies.

Many thanks to @redguardtoo for making it possible!

P.S. code is not battle tested and far from being ideal, but it works for my
needs.