[emacs] prepare functions for wine region migration
The plan is to move each appellation to a separate note in org-roam.
`wine/migrate-region` ensures that the associated country and geographical
region exists in org-roam and creates an appellation if needed. Then it links a
wine entry to org-roam appellation and geographical region. As the last step it
removes org-brain appellation entry.