[emacs] helpers for reading props from description list
PROPERTIES drawer does not support links by design. For notes it means that
references are not detected from PROPERTIES drawer (basically, because elements
API doesn't see them as links). In addition, PROPERTIES drawer should contain
mechanical information that is not intended to be always visible.

So I could not find other approach rather than use a description list to define
properties that I want to query (and in future set). This commit provides two
functions for end users:

- `+org-notes-meta-get` to get the value of desired property,
- `+org-notes-meta-get-list` to get list of value of desired property (for cases
  when property is encountered multiple times).