[emacs] dynamically build agenda files list
Instead of searching for TODOs in ALL org-mode files (around 1700 right now),
search only in files marked as Projects. Ideally it should not use
`org-roam-db-query` directly, but rather some `org-notes` API and it also should
not use hard-coded inbox files, but this solution is quick and gives what I
wanted it to achieve - 25seconds -> <2 seconds for building agenda.

It is possible also to avoid recalculating this value every time agenda is
called, using some hooks from org-notes.

This solution has a drawback - required files might be missed if they are not
properly marked. This will be improved separately by ensuring tags on-save.