(nixos) setup gnupg
In general, it's a pain in the ass to change default homedir of gnupg.
I am pretty sure that's me doing something completely wrong, but I
wasted enough time on this.

Since I have zero manually written files in $XDG_CONFIG_HOME/gnupg,
there is no reason to stick with it for me (aside for less clutter in
$HOME, but I prefer $XDG_CONFIG_HOME to be clean as I sync it).

So... here we are