(system) use xsession instead of xinit
It's not clear from this commit why as the real reason is hidden in
private scripts.

In short, because of zoom-us that I have to use for work, I am forced
to use desktop environment and display manager, otherwise zoom-us
doesn't function properly - it either crashes or doesn't show any
video.

So my setup is:

- os: Arch Linux
- de: xfce with most of the stuff disabled
- dm: LightDM

Custom stuff is configured via xsession file