emacs:barberry-theme: tune and add colors
Aside from tuning existing colors into my linking, this commit has
three important changes.

First, it adds a color for 'successful' type of information. IMO using
salient color for this purpose is not helpful.

Second, it adds inverse version of most colors, because (and this is
one of the biggest issues with NANO theme) not all colors are good in
both variants. Basically, I am picking one color and using different
shades of the same color.

Third, it adds extra highlight colors for cases when we need
background to highlight something. Mostly used in VCS related stuff
for now.