emacs:publicatorg: flatten deps
The goal of this change is to simplify deps manipulations. One of the ways to achieve this is to allow rules to have
multiple outputs, where every output can have zero or more dependencies on other outputs. And this required also to
introduce separate compilers.

For example, wine post rule outputs a single org file, and multiple images. The org file output depends on ratings,
producer and it's images. The images have zero dependencies. The org file is compiled using specific compiler for wine
posts and images are compiled with compiler for images. You can see more example in the build-rules.el of
d12frosted/barberry.io.

While working on this I wanted to get rid of plist files and introduce better structure by using defstructs.